Change project folder, treatment for equal name
If the user chooses to set a container name, now it becomes part of
folder root. This helps to make te final output content be more self
explanatory, as sanning the case where an existing folder already
exists.
Also in case of a folder already exists, a better message output
is provided. The problem is treated for the specific case.